refactor layout and add raw ui for company account setting
