+++ app/(protected)/layout.tsx
@@ -1,25 +1,5 @@
 import React from 'react'
-import type { Metadata } from 'next'
-import { AIPopup, Navbar, NotificationSnackBar } from '@/components'
-
-import localFont from 'next/font/local'
-import './globals.css'
-
-const geistSans = localFont({
-  src: './fonts/GeistVF.woff',
-  variable: '--font-geist-sans',
-  weight: '100 900'
-})
-const geistMono = localFont({
-  src: './fonts/GeistMonoVF.woff',
-  variable: '--font-geist-mono',
-  weight: '100 900'
-})
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app'
-}
+import { AIPopup, Navbar } from '@/components'
 
 export default function RootLayout({
   children
@@ -27,13 +7,10 @@ export default function RootLayout({
   children: React.ReactNode
 }>) {
   return (
-    <html lang='en'>
-      <body className={`${geistSans.variable} ${geistMono.variable} antialiased bg-background`}>
-        <NotificationSnackBar />
-        <Navbar />
-        {children}
-        <AIPopup />
-      </body>
-    </html>
+    <>
+      <Navbar />
+      {children}
+      <AIPopup />
+    </>
   )
 }
