+++ app/layout.tsx
@@ -1,7 +1,20 @@
 import React from 'react'
 import type { Metadata } from 'next'
+import { NotificationSnackBar } from '@/components'
+
 import localFont from 'next/font/local'
-import { Navbar, NotificationSnackBar } from '@/components'
+import './globals.css'
+
+const geistSans = localFont({
+  src: './fonts/GeistVF.woff',
+  variable: '--font-geist-sans',
+  weight: '100 900'
+})
+const geistMono = localFont({
+  src: './fonts/GeistMonoVF.woff',
+  variable: '--font-geist-mono',
+  weight: '100 900'
+})
 
 export const metadata: Metadata = {
   title: 'Create Next App',
@@ -15,7 +28,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang='en'>
-      <body className={`antialiased bg-background`}>
+      <body className={`${geistSans.variable} ${geistMono.variable} antialiased bg-background`}>
         <NotificationSnackBar />
         {children}
       </body>
